Refactored subclasses of PlayerTextures to a factory method of itself and renamed to TexturesHolder.h since it can be used more generally
