--- include/entity/player/sprites/PlayerIdle.h
@@ -1,28 +1,0 @@
-//
-// Created by Agamjeet Singh on 25/07/25.
-//
-
-#ifndef PLAYERIDLE_H
-#define PLAYERIDLE_H
-#include "PlayerTextures.h"
-#include <SFML/Graphics.hpp>
-
-
-class PlayerIdle final : public PlayerTextures {
-public:
-    PlayerIdle() {
-        for (int i = 0; i < 9; i++) {
-            sf::Texture texture;
-            assert(texture.loadFromFile("../assets/player/idle/idle0" + std::to_string(i) + ".png"));
-            textures.push_back(texture);
-        }
-
-        for (int i = 0; i < 9; i++) {
-            intervals.push_back(0.1);
-        }
-    }
-};
-
-
-
-#endif //PLAYERIDLE_H
